Update LM generate recipe
